# Chapter09/Chapter09.Service/Models/WeatherForecast.cs
using System;

namespace Chapter09.Service.Models
{
    public class WeatherForecast
    {
        public DateTime Date { get; set; }

        public int TemperatureC { get; set; }

        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);

        public string Summary { get; set; }
    }
}